requirements.txt
streamlit
face-recognition
Pillow
numpy
pandas
cmake
scikit-build
dlib==19.24.1
